moved logic to lib module and added benchmarks
