+++ Cargo.toml
@@ -11,3 +11,10 @@ term-table = "1.3.2"
 colored = "2.0.0"
 clap = { version="4.2.4", features=["derive"]}
 serde_json = "1.0.96"
+
+[dev-dependencies]
+criterion = "0.4.0"
+
+[[bench]]
+name = "benches"
+harness = false
